device_driver_mir::model

Enum EnumGenerationStyle

pub enum EnumGenerationStyle {
    Fallible,
    InfallibleWithinRange,
    Fallback,
}

Variants

Fallible

Not all basetype values can be converted to a variant

InfallibleWithinRange

All bitpatterns within bits 0..size-bits are covered. The general interface is fallible, but this special knowledge can be used for safety guarantees

Fallback

There’s a fallback, so it’s always safe

Implementations

impl EnumGenerationStyle

pub fn is_fallible(&self) -> bool

Returns true if the enum generation style is Fallible.