device_driver_mir::model

Enum Object

pub enum Object {
    Device(Device),
    Block(Block),
    Register(Register),
    Command(Command),
    Buffer(Buffer),
    FieldSet(FieldSet),
    Enum(Enum),
    Extern(Extern),
    Field(Field),
}

Variants

Device(Device)

Block(Block)

Register(Register)

Command(Command)

Buffer(Buffer)

FieldSet(FieldSet)

Enum(Enum)

Extern(Extern)

Field(Field)

Implementations

impl Object

pub fn device_config(&self) -> Option<&DeviceConfig>

pub fn child_objects_mut(&mut self) -> &mut [Object]

pub fn child_objects_vec(&mut self) -> Option<&mut Vec<Object>>

pub fn child_objects(&self) -> &[Object]

pub fn name_mut(&mut self) -> &mut Identifier<RuntimeType>

Get a mutable reference to the name of the specific object

pub fn name(&self) -> &Identifier<RuntimeType>

Get a reference to the name of the specific object

pub fn name_span(&self) -> Span

Get the span of the name of the object

pub fn address(&self) -> Option<Spanned<i128>>

Return the address if it is specified.

pub fn repeat(&self) -> Option<&Repeat>

Return the repeat value if it exists

pub fn repeat_mut(&mut self) -> Option<&mut Repeat>

Return the repeat value if it exists

pub fn as_field_set(&self) -> Option<&FieldSet>

pub fn as_field_set_mut(&mut self) -> Option<&mut FieldSet>

pub fn as_enum(&self) -> Option<&Enum>

pub fn allow_address_overlap(&self) -> bool

pub fn span(&self) -> Span

The span of the entire object

pub fn node_type(&self) -> NodeType

pub fn fieldset_refs(&self) -> Vec<Spanned<IdentifierRef<Type>>>

Get the fieldset refs of the object. Only returns non-zero for registers and commands

pub fn properties_span(&self) -> Option<Span>